refactor(client/config): move the client and config to their own modules
This should help in making the project more navigable and organized.

This refactor has caused some minor changes in functionality. These are
(for the most part) marked with comments and will be addressed properly
before the next release tag.

This is only a partial refactor as there are still many things that need
to be re-organized.