refactor(client): partial refactor to create a client struct
This is by far not done, but I am splitting this into different commits,
since the next step will require the replacement of rouille, since it's
implementation of websockets will not work for the new architecture.

Besides this rouille has been causing some pains for a while now, so a
replacement would be good anyway.